# Optimize assignments