--- a/lib/WorseReflection/Core/Inference/Resolver/MemberAccess/NodeContextFromMemberAccess.php
+++ b/lib/WorseReflection/Core/Inference/Resolver/MemberAccess/NodeContextFromMemberAccess.php
@@ -254,8 +254,8 @@ class NodeContextFromMemberAccess
         }
 
         $variable = $frame->properties()
-            ->lessThanOrEqualTo($position)
             ->byName($propertyName)
+            ->lessThanOrEqualTo($position)
             ->lastOrNull();
 
         if (null === $variable) {