--- a/phpstan-baseline.neon
+++ b/phpstan-baseline.neon
@@ -691,21 +691,6 @@ parameters:
 			path: lib/CodeTransform/Adapter/WorseReflection/Transformer/ImplementContracts.php
 
 		-
-			message: "#^Cannot access property \\$children on Microsoft\\\\PhpParser\\\\Node\\\\DelimitedList\\\\NamespaceUseGroupClauseList\\|null\\.$#"
-			count: 1
-			path: lib/CodeTransform/Adapter/WorseReflection/Transformer/RemoveUnusedImportsTransformer.php
-
-		-
-			message: "#^Cannot call method getEndPosition\\(\\) on Microsoft\\\\PhpParser\\\\Node\\|null\\.$#"
-			count: 2
-			path: lib/CodeTransform/Adapter/WorseReflection/Transformer/RemoveUnusedImportsTransformer.php
-
-		-
-			message: "#^Cannot call method getStartPosition\\(\\) on Microsoft\\\\PhpParser\\\\Node\\|null\\.$#"
-			count: 2
-			path: lib/CodeTransform/Adapter/WorseReflection/Transformer/RemoveUnusedImportsTransformer.php
-
-		-
 			message: "#^Method Phpactor\\\\CodeTransform\\\\CodeTransform\\:\\:transform\\(\\) has parameter \\$transformations with no value type specified in iterable type array\\.$#"
 			count: 1
 			path: lib/CodeTransform/CodeTransform.php
@@ -1452,7 +1437,7 @@ parameters:
 
 		-
 			message: "#^Parameter \\#1 \\$type of method Phpactor\\\\DocblockParser\\\\Parser\\:\\:createTypeFromToken\\(\\) expects Phpactor\\\\DocblockParser\\\\Ast\\\\Token, Phpactor\\\\DocblockParser\\\\Ast\\\\Token\\|null given\\.$#"
-			count: 4
+			count: 3
 			path: lib/DocblockParser/Parser.php
 
 		-
@@ -3981,16 +3966,6 @@ parameters:
 			path: lib/Extension/LanguageServerCodeTransform/CodeAction/CreateClassProvider.php
 
 		-
-			message: "#^Cannot cast mixed to string\\.$#"
-			count: 1
-			path: lib/Extension/LanguageServerCodeTransform/CodeAction/CreateUnresolvableClassProvider.php
-
-		-
-			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\CodeAction\\\\CreateUnresolvableClassProvider\\:\\:provideActionsFor\\(\\) should return Amp\\\\Promise\\<array\\<Phpactor\\\\LanguageServerProtocol\\\\CodeAction\\>\\> but returns Amp\\\\Promise\\<mixed\\>\\.$#"
-			count: 1
-			path: lib/Extension/LanguageServerCodeTransform/CodeAction/CreateUnresolvableClassProvider.php
-
-		-
 			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\CodeAction\\\\ExtractConstantProvider\\:\\:provideActionsFor\\(\\) should return Amp\\\\Promise\\<array\\<Phpactor\\\\LanguageServerProtocol\\\\CodeAction\\>\\> but returns Amp\\\\Promise\\<mixed\\>\\.$#"
 			count: 1
 			path: lib/Extension/LanguageServerCodeTransform/CodeAction/ExtractConstantProvider.php
@@ -4011,17 +3986,7 @@ parameters:
 			path: lib/Extension/LanguageServerCodeTransform/CodeAction/GenerateDecoratorProvider.php
 
 		-
-			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\CodeAction\\\\GenerateMethodProvider\\:\\:provideActionsFor\\(\\) should return Amp\\\\Promise\\<array\\<Phpactor\\\\LanguageServerProtocol\\\\CodeAction\\>\\> but returns Amp\\\\Promise\\<mixed\\>\\.$#"
-			count: 1
-			path: lib/Extension/LanguageServerCodeTransform/CodeAction/GenerateMethodProvider.php
-
-		-
 			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\CodeAction\\\\ImportNameProvider\\:\\:diagnosticsFromUnresolvedName\\(\\) return type has no value type specified in iterable type array\\.$#"
-			count: 1
-			path: lib/Extension/LanguageServerCodeTransform/CodeAction/ImportNameProvider.php
-
-		-
-			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\CodeAction\\\\ImportNameProvider\\:\\:provideDiagnostics\\(\\) should return Amp\\\\Promise\\<array\\<Phpactor\\\\LanguageServerProtocol\\\\Diagnostic\\>\\> but returns Amp\\\\Promise\\<mixed\\>\\.$#"
 			count: 1
 			path: lib/Extension/LanguageServerCodeTransform/CodeAction/ImportNameProvider.php
 
@@ -4034,11 +3999,6 @@ parameters:
 			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\CodeAction\\\\ReplaceQualifierWithImportProvider\\:\\:provideActionsFor\\(\\) should return Amp\\\\Promise\\<array\\<Phpactor\\\\LanguageServerProtocol\\\\CodeAction\\>\\> but returns Amp\\\\Promise\\<mixed\\>\\.$#"
 			count: 1
 			path: lib/Extension/LanguageServerCodeTransform/CodeAction/ReplaceQualifierWithImportProvider.php
-
-		-
-			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\CodeAction\\\\TransformerCodeActionPovider\\:\\:provideActionsFor\\(\\) should return Amp\\\\Promise\\<array\\<Phpactor\\\\LanguageServerProtocol\\\\CodeAction\\>\\> but returns Amp\\\\Promise\\<mixed\\>\\.$#"
-			count: 1
-			path: lib/Extension/LanguageServerCodeTransform/CodeAction/TransformerCodeActionPovider.php
 
 		-
 			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\Converter\\\\DiagnosticsConverter\\:\\:toLspDiagnostics\\(\\) return type has no value type specified in iterable type array\\.$#"
@@ -4124,11 +4084,6 @@ parameters:
 			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\LspCommand\\\\ReplaceQualifierWithImportCommand\\:\\:__invoke\\(\\) should return Amp\\\\Promise\\<Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\LspCommand\\\\ApplyWorkspaceEditResult\\|null\\> but returns Amp\\\\Promise\\<Phpactor\\\\LanguageServerProtocol\\\\ApplyWorkspaceEditResult\\>\\.$#"
 			count: 1
 			path: lib/Extension/LanguageServerCodeTransform/LspCommand/ReplaceQualifierWithImportCommand.php
-
-		-
-			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\LspCommand\\\\TransformCommand\\:\\:__invoke\\(\\) return type with generic interface Amp\\\\Promise does not specify its types\\: TValue$#"
-			count: 1
-			path: lib/Extension/LanguageServerCodeTransform/LspCommand/TransformCommand.php
 
 		-
 			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerCodeTransform\\\\Model\\\\NameImport\\\\NameImporterResult\\:\\:__construct\\(\\) has parameter \\$textEdits with no value type specified in iterable type array\\.$#"
@@ -4731,21 +4686,6 @@ parameters:
 			path: lib/Extension/LanguageServerPhpstan/Tests/Provider/PhpstanDiagnosticProviderTest.php
 
 		-
-			message: "#^Cannot call method resolve\\(\\) on mixed\\.$#"
-			count: 2
-			path: lib/Extension/LanguageServerPsalm/LanguageServerPsalmExtension.php
-
-		-
-			message: "#^Parameter \\#2 \\$shouldShowInfo of class Phpactor\\\\Extension\\\\LanguageServerPsalm\\\\Model\\\\PsalmConfig constructor expects bool, mixed given\\.$#"
-			count: 1
-			path: lib/Extension/LanguageServerPsalm/LanguageServerPsalmExtension.php
-
-		-
-			message: "#^Parameter \\#3 \\$useCache of class Phpactor\\\\Extension\\\\LanguageServerPsalm\\\\Model\\\\PsalmConfig constructor expects bool, mixed given\\.$#"
-			count: 1
-			path: lib/Extension/LanguageServerPsalm/LanguageServerPsalmExtension.php
-
-		-
 			message: "#^Cannot access offset 'column_from' on mixed\\.$#"
 			count: 1
 			path: lib/Extension/LanguageServerPsalm/Model/DiagnosticsParser.php
@@ -4794,11 +4734,6 @@ parameters:
 			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerPsalm\\\\Tests\\\\Model\\\\DiagnosticsParserTest\\:\\:provideParse\\(\\) has no return type specified\\.$#"
 			count: 1
 			path: lib/Extension/LanguageServerPsalm/Tests/Model/DiagnosticsParserTest.php
-
-		-
-			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerPsalm\\\\Tests\\\\Model\\\\PsalmProcessTest\\:\\:testLint\\(\\) has parameter \\$expectedDiagnostics with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: lib/Extension/LanguageServerPsalm/Tests/Model/PsalmProcessTest.php
 
 		-
 			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerReferenceFinder\\\\Handler\\\\GotoImplementationHandler\\:\\:gotoImplementation\\(\\) return type with generic interface Amp\\\\Promise does not specify its types\\: TValue$#"
@@ -5174,11 +5109,6 @@ parameters:
 			message: "#^Parameter \\#1 \\$workspace of class Phpactor\\\\Extension\\\\LanguageServerSymbolProvider\\\\Handler\\\\DocumentSymbolProviderHandler constructor expects Phpactor\\\\LanguageServer\\\\Core\\\\Workspace\\\\Workspace, mixed given\\.$#"
 			count: 1
 			path: lib/Extension/LanguageServerSymbolProvider/LanguageServerSymbolProviderExtension.php
-
-		-
-			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerWorseReflection\\\\DiagnosticProvider\\\\WorseDiagnosticProvider\\:\\:provideDiagnostics\\(\\) should return Amp\\\\Promise\\<array\\<Phpactor\\\\LanguageServerProtocol\\\\Diagnostic\\>\\> but returns Amp\\\\Promise\\<mixed\\>\\.$#"
-			count: 1
-			path: lib/Extension/LanguageServerWorseReflection/DiagnosticProvider/WorseDiagnosticProvider.php
 
 		-
 			message: "#^Method Phpactor\\\\Extension\\\\LanguageServerWorseReflection\\\\Tests\\\\Benchmark\\\\WorkspaceIndexBench\\:\\:benchUpdate\\(\\) has parameter \\$params with no value type specified in iterable type array\\.$#"
@@ -6667,11 +6597,6 @@ parameters:
 
 		-
 			message: "#^Parameter \\#1 \\$reflector of class Phpactor\\\\Indexer\\\\Adapter\\\\ReferenceFinder\\\\Util\\\\ContainerTypeResolver constructor expects Phpactor\\\\WorseReflection\\\\Core\\\\Reflector\\\\ClassReflector, mixed given\\.$#"
-			count: 1
-			path: lib/Indexer/Extension/IndexerExtension.php
-
-		-
-			message: "#^Parameter \\#1 \\$reflector of class Phpactor\\\\Indexer\\\\Adapter\\\\Worse\\\\WorseRecordReferenceEnhancer constructor expects Phpactor\\\\WorseReflection\\\\Core\\\\Reflector\\\\SourceCodeReflector, mixed given\\.$#"
 			count: 1
 			path: lib/Indexer/Extension/IndexerExtension.php