Further reduces the PHPStan benchmark from 11s to 2.6s

This is quite specific however - the time to analyse the WorseReflection package is only reduced from 22s to 21s